refactor: common component
