+++ packages/common/i18n/index.js
@@ -1,0 +1,9 @@
+import i18n, { i18nKeyMaps } from '@opentiny/tiny-engine-controller/js/i18n'
+
+import enUs from './en-us.json'
+import zhCn from './zh-cn.json'
+
+const { mergeLocaleMessage } = i18n.global
+
+mergeLocaleMessage(i18nKeyMaps.enUS, enUs)
+mergeLocaleMessage(i18nKeyMaps.zhCN, zhCn)
